- Add a precheck for the Debian ISO - Change Qemu VM name to bidms.qcow2 so the image file is named right - Add a $HOME/.aptproxy file - Add bidms user to docker group - Install the bidms_nw docker network - Set up a crontab to prune unused docker images

+++ build-qemu.sh
@@ -27,6 +27,12 @@
 # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 # 
 
+./precheck.sh
+if [ $? != 0 ]; then
+  echo "Prerequisites failed"
+  exit 1
+fi
+
 if [ -e output-qemu ]; then
   echo "Deleting previous build in the output-qemu directory"
   rm -rf output-qemu
